MidiPlayer: Update version to 1.0.1 final
Change the BAlert about box into a BAboutWindow.
Add Authors based on commit logs.

Unfortunately the description contained:
"Haiku MIDI Player 1.0.0 beta\n\n" which means that it
is going to get flagged for re-translation, sorry! This was
bound to happen when the version number changed,
on the bright side it shouldn't happen again in the future
since the Name and version number are separate.