MidiPlayer: Rework layout
#9882 comment:8 suggested applying a similar fix
to MidiPlayer as the one to Backgrounds, which makes
the menu fields variable width and isn't affected by the
check box.

In addition this replaces BGroupLayoutBuilder and
BGridLayoutBuilder with BLayoutBuilder templates and
replaces hardcoded spacing units with B_USE_ spacing
constants.

I also put the Volume slider on its own line and tightened
up the spacing around the scope view.

picture:
http://31.media.tumblr.com/74d93b806033d9adee83eeb2adead48d/tumblr_n7v06onPdl1r0f0hfo1_250.png