Add unbootstrap, a script to remove _bootstrap from packages.
This needs to be run under haiku as it uses the package tool. It can be
used to remove the _bootstrap tagging from bootstrap packages and then
upload them to the repos as normal packages. The goal is to make
non-bootstrap builds work even for archs where the bootstrap isn't
working yet.

Packages built that way should be replaced as soon as better ones are
obtained by the proper bootstrapping process.