libbind: allow underscores in domain names
As discussed in http://www.freelists.org/post/haiku-development/Proposal-adding-nochecknames-to-resolvconf-by-default

The underscore is handled like the hyphen, that is, it can only be in
the middle of a word. "a_b.com" is valid, but "a_.com" and "_a.com" are
not.

Patch is from http://lists.freebsd.org/pipermail/freebsd-arch/2003-March/000071.html