pc_serial: fix KDL on close
It's working \o/
Just need to figure out why it waits for \n to transmit.

We now only destroy the cookies in the free hook.

Mention pending DPC as soon as when inside the irq handler.

Also limit the number of loops in the IRQ handler to avoid busy looping.
It only happened when testing with normal priority but you never know...