tools: data_to_source: fix types of generated variables
Both data array and its size should be constant so that GCC can do
as much as possible at compile time and it is safe to use them in
static assertions.

This patch also changes the type of size constant to size_t which
more appropriate and uses sizeof() to determine the size of the
array.

Fixes build breakage introduced in e547662664d88e5ee79048bd00ad8eefa45e7074.

Signed-off-by: Paweł Dziepak <pdziepak@quarnos.org>