fix for older Wacom tablet
Signed-off-by: Adrien Destugues <pulkomandy@pulkomandy.tk>

This spent enough years waiting on the bugtracker for someone willing to
test it. The style looks mostly ok, so let's have it in trunk and see if
it breaks someone's tablet.

The patch restores the older behavior found in early versions of the
driver published for BeOS. It does so only for the "intuos" tablets, so
newer devices won't be affected.

ticket : #4847