Chart: clear view when switiching to BDirectWindow
Even when BDirectWindow is used, the back buffer data can be copied to
the front buffer. This happens when resizing or moving the window, and
when moving the cursor above it.

Since Charts does not touch the whole window when redrawing (it only
erases previous stars and draws new ones) this led to the last frame
drawn on the back buffer to stay visible until stars erased it.

Clearing the back buffer view with the current "sapce color" makes this
much less visible.

Fixes #96.