Revert update to ffmpeg 0.11.5 for gcc2.
More testing shown that it crashes when decoding some videos which
work with 0.10 (mostly h263 and h264). Gcc4 versions does not seem to
have a problem, so we can keep the update at least there (WebKit and
other gcc4 compiled apps will use it).

Note that the changes in the media plugin are not reversed: they are
compatible with both ffmpeg versions.