GIF and WebP translators: force layout of parent window
Translators are now allowed to use the layout API, but this doesn't work
well when they are used in a non-layout aware window (the view ends up
with a size of 0x0 or some arbitrary size set with
SetExplicitPreferredSize).

To avoid this, detect the case where the translator settings are the
single child of a non-layout window, and force the window to become
layouted.

Fixes #7754.